requirements.txt
langchain>=0.0.335
python-dotenv>=1.0.0
openai>=1.3.5
chroma-hnswlib>=0.7.3
chromadb>=0.4.15
tqdm>=4.66.1
pandas>=2.1.3
numpy>=1.26.2
faiss-cpu>=1.7.4
torch>=2.1.1
transformers>=4.35.2
sentence-transformers>=2.2.2